Fatkhur

+++ list-kategori-produk.php
@@ -89,7 +89,7 @@
 											<div class="dropdown-menu dropdown-menu-right">
 												<a class="dropdown-item" href="#"><i class="fa fa-eye"></i> View</a>
 												<a class="dropdown-item" href="edit-product-category.php?edit=<?=$row['id']?>&table=<?= $t++ ;?>"><i class="fa fa-pencil"></i> Edit</a>
-												<a class="dropdown-item" href="list-kategori-produk.php?delete=<?=$row['id']?>"><i class="fa fa-trash"></i> Delete</a>
+												<a class="dropdown-item" href="list-kategori-produk.php?delete=<?=$row['id']?>" onclick="return confirm('Are you sure you want to delete?')"><i class="fa fa-trash"></i> Delete</a>
 											</div>
 										</div>
 									</td>
